refactor: move toolbar from OscdSclEditor to OscdAceEditor
OscdAceEditor now provides undo/redo, collapse/expand, search,
format-XML, and settings toolbar buttons behind a `toolbar` property.

OscdSclEditor's render is simplified to a single
`<oscd-ace-editor ?toolbar=${this.toolbar}>` delegation.
